extracted NVSHandle to new header file
