+++ doc/en/nvs.rst
@@ -6,7 +6,14 @@ NVS
 
 The `NVS` component provides a simple class representing an NVS controller. 
 
-Code examples for the task API are provided in the `nvs` example folder.
+NVSHandle
+---------
+
+The `NVSHandle` class manages individual NVS storage handles, allowing for scoped
+control over specific NVS namespaces. It simplifies operations like reading,
+writing, and committing key-value pairs within these namespaces. 
+
+Code examples for the NVS and NVSHandle API are provided in the `nvs` example folder.
 
 .. ---------------------------- API Reference ----------------------------------
 
@@ -14,3 +21,4 @@ API Reference
 -------------
 
 .. include-build-file:: inc/nvs.inc
+.. include-build-file:: inc/nvs_handle_espp.inc
